rename: albumListSection으로 컴포넌트명 변경
