+++ src/pages/admin/albums/index.page.tsx
@@ -1,4 +1,4 @@
-import AlbumListSection from "@/components/album/AlbumsListSection/AlbumsListSection";
+import AlbumListSection from "@/components/album/AlbumListSection/AlbumListSection";
 
 const AdminAlbumListPage = () => {
   return (
